+++ lib/i18n/tasks/data/tree/traversal.rb
@@ -64,8 +64,10 @@ module I18n::Tasks::Data::Tree
       keys(opts).map { |key, node| [key, node.value] }
     end
 
-    def root_key_values
-      keys(root: false).map { |key, node| [node.root.key, key, node.value]}
+    def root_key_values(sort = false)
+      result = keys(root: false).map { |key, node| [node.root.key, key, node.value]}
+      result.sort! { |a, b| a[0] != b[0] ? a[0] <=> b[0] : a[1] <=> b[1] } if sort
+      result
     end
 
     #-- modify / derive